Add files via upload
The Royal Mint Café is a dining establishment specializing in providing high-quality food and beverages. It focuses on delivering a seamless experience by offering services such as cash-on-delivery orders, phone-based bookings, and prompt food delivery by dedicated staff. The café aims to cater to a diverse customer base with a user-friendly interface for menu browsing and efficient service, ensuring customer satisfaction and convenience.